+++ Backend/server/Controllers/activityController.js
@@ -532,9 +532,17 @@ const cardPaymentSuccess = async (req, res) => {
 };
 
 const cancelActivityBooking = async (req, res) => {
+
   const { bookingId } = req.params; // Get the booking ID from the request params
 
   try {
+    const authHeader = req.header("Authorization");
+    if (!authHeader) {
+      return res.status(401).json({ message: "Authorization header missing" });
+    }
+    const token = authHeader.replace("Bearer ", "");
+    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+
     const booking = await Booking.findById(bookingId);
 
     if (!booking) {
@@ -558,6 +566,12 @@ const cancelActivityBooking = async (req, res) => {
         message: `Cannot cancel booking less than 48 hours before the event, hoursDifference=${hoursDifference}`,
       });
     }
+    const tourist = await touristModel.findOne({ _id: decoded.id });
+    if (!tourist) {
+      return res.status(404).json({ message: "Tourist not found" });
+    }
+    tourist.wallet += activity.price;
+    await tourist.save();
 
     // If it's more than 48 hours, proceed to cancel the booking
     await Booking.deleteOne({ _id: bookingId });
